Aryesh

+++ controller/user/DeleteUserCtrl.js
@@ -4,7 +4,18 @@ const validId = require("../../utils/isValid");
 
 const DeleteUserCtrl = expressAsyncHandler(async (req, res) => {
     const id = req.user._id;
-    validId(id);
+    // const email = req.body.email;
+    const pass = req.body.password;
+
+    validId(myId)
+
+    const user = await User.findById(myId)
+
+    // if(user.email !== email) throw new Error("Incorrect Email")
+
+    const checkPass = await user.CheckPassword(pass)
+
+    if(!checkPass) throw new Error("Incorrect Password")
 
     try {
         const deleteUser = await User.findByIdAndDelete(id);
